2008-04-09  H.J. Lu  <hongjiu.lu@intel.com>
	PR middle-end/35897
	* dse.c (store_info): Use long long on positions_needed.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/ix86@134160 138bc75d-0d04-0410-961f-82ee72b054a4

+++ avx/gcc/dse.c
@@ -229,7 +229,8 @@ struct store_info
   /* An bitmask as wide as the number of bytes in the word that
      contains a 1 if the byte may be needed.  The store is unused if
      all of the bits are 0.  */
-  long positions_needed;
+  /* FIXME: A real bitmap should be used.  */
+  long long positions_needed;
 
   /* The next store info for this insn.  */
   struct store_info *next;
